requirements.txt
bentoml==1.1.10
loggerFactory==0.0.6
boto3==1.20.54
gast==0.4.0
grpcio==1.50.0
gunicorn==20.0.4
Keras-Preprocessing==1.1.2
libclang==14.0.6
linc-detector==0.1.3
numpy==1.23.3
pandas==2.0.3
pydantic==2.5.1
torch==1.10.1
torchvision==0.11.2
schema==0.7.5